relocate code and add comments
