+++ document.go
@@ -91,3 +91,26 @@ type TokenizableSpatialField interface {
 	// to override the spatial token generations during the analysis phase.
 	SetSpatialAnalyzerPlugin(SpatialAnalyzerPlugin)
 }
+
+// SynonymField represents a field that contains a list of synonyms for a set of terms.
+// Each SynonymField is generated from a single synonym definition, and its name corresponds
+// to the synonym source to which the synonym definition belongs.
+type SynonymField interface {
+	Field
+	// IterateSynonyms iterates over the synonyms for the term in the field.
+	// The provided visitor function is called with each term and its corresponding synonyms.
+	IterateSynonyms(visitor func(term string, synonyms []string))
+}
+
+// SynonymFieldVisitor is a function type used to visit a SynonymField within a document.
+type SynonymFieldVisitor func(SynonymField)
+
+// SynonymDocument represents a special type of document that contains synonym fields.
+// Each SynonymField is a field with a list of synonyms for a set of terms.
+// These fields are derived from synonym definitions, and their names correspond to the synonym sources.
+type SynonymDocument interface {
+	Document
+	// VisitSynonymFields allows iteration over all synonym fields in the document.
+	// The provided visitor function is called for each synonym field.
+	VisitSynonymFields(visitor SynonymFieldVisitor)
+}
